feat(protocols): move nvext handling to llm
