+++ lib/llm/src/protocols/unified.rs
@@ -40,9 +40,9 @@ use serde::{Deserialize, Serialize};
 use crate::preprocessor::media::MediaDecoder;
 use dynamo_renderer::{OAIChatLikeRequest, TextInput};
 
+use crate::protocols::common::extensions::{NvExt, NvExtProvider};
 use crate::protocols::openai::chat_completions::NvCreateChatCompletionRequest;
 use crate::protocols::openai::common_ext::{CommonExt, CommonExtProvider};
-use crate::protocols::openai::nvext::{NvExt, NvExtProvider};
 use crate::protocols::openai::{
     OpenAIOutputOptionsProvider, OpenAISamplingOptionsProvider, OpenAIStopConditionsProvider,
 };
@@ -572,6 +572,7 @@ mod tests {
                     content: "Hello".to_string(),
                 },
             }],
+            nvext: None,
             system: None,
             temperature: Some(0.7),
             top_p: None,
